Allow MessageContext to accept custom macros
